+++ packages/kap-server/src/routes/plugins.ts
@@ -11,13 +11,15 @@
  * catalogs and the capability shelf-install hook). The marketplace catalog is
  * read on demand from the configured location (`pluginMarketplaceUrl` server
  * option, env `KIMI_CODE_PLUGIN_MARKETPLACE_URL`, default the production
- * catalog; plain paths and `file://` URLs read from disk like the CLI loader)
- * and merged with the live install state — install status is always detected
- * from the local records, never from the catalog. Catalog-relative sources
- * (`./official/*.zip`) resolve against the catalog location so the returned
- * `source` is directly installable. Entries without a `version` get one from
- * a GitHub ref tail or the bare repo's latest release (CLI parity), which is
- * what drives `updateAvailable`.
+ * catalog) through the shared `app/plugin/marketplace` client — catalog
+ * reading, the lenient entry normalization, source resolution, and version
+ * derivation all live there (one implementation, consumed by the CLI too).
+ * When the location is the built-in default, a failed read falls back to the
+ * source checkout's own catalog (offline dev); an explicitly configured
+ * catalog fails hard. The route merges the entries with the live install
+ * state — install status is always detected from the local records, never
+ * from the catalog — and marks capability wiring rows with `capabilityId` so
+ * clients route them through `/capabilities/{id}:install`.
  *
  * **Action suffix**: `:enable` / `:disable` / `:remove` via `parseActionSuffix`
  * (bare ids rejected).
@@ -30,20 +32,22 @@
  *   - other errors                 → `50001` via the global error handler
  */
 
+import { stat } from 'node:fs/promises';
+import { resolve } from 'node:path';
+
 import {
+  computeUpdateStatus,
   ErrorCodes as DomainErrorCodes,
   IPluginService,
   PluginErrors,
   isError2,
+  parsePluginMarketplace,
+  readPluginMarketplace,
+  withLatestVersions,
+  type MarketplaceLocation,
+  type PluginMarketplace,
   type Scope,
 } from '@moonshot-ai/agent-core-v2';
-import { existsSync } from 'node:fs';
-import { readFile } from 'node:fs/promises';
-import { homedir } from 'node:os';
-import { dirname, isAbsolute, join, resolve } from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-import { gt, valid } from 'semver';
 import { z } from 'zod';
 
 import { errEnvelope, okEnvelope } from '../envelope';
@@ -101,254 +105,21 @@ const CAPABILITY_ROW_IDS: Readonly<
 
 const MARKETPLACE_FETCH_TIMEOUT_MS = 10_000;
 
-// Custom catalogs accepted by the CLI may carry the source under the legacy
-// `url` / `downloadUrl` aliases — normalize before validating so a catalog
-// that works in the CLI works here too.
-const rawMarketplaceEntrySchema = z.preprocess(
-  (value) => {
-    if (typeof value !== 'object' || value === null) return value;
-    const record = value as Record<string, unknown>;
-    // CLI stringField semantics: non-string or blank counts as missing, and
-    // the first valid of source / url / downloadUrl wins (trimmed).
-    const pick = (v: unknown) =>
-      typeof v === 'string' && v.trim().length > 0 ? v.trim() : undefined;
-    const normalized: Record<string, unknown> = { ...record };
-    // The id keys the install-state join — trim it like the CLI's
-    // requiredString; a blank id drops out so the schema rejects the entry.
-    const id = pick(record['id']);
-    if (id === undefined) delete normalized['id'];
-    else normalized['id'] = id;
-    // Metadata: blank reads as missing, and the CLI parser's aliases are
-    // honored (name / shortDescription / websiteURL).
-    const metadataAliases = [
-      ['displayName', 'name'],
-      ['description', 'shortDescription'],
-      ['homepage', 'websiteURL'],
-    ] as const;
-    for (const [field, alias] of metadataAliases) {
-      const value = pick(record[field]) ?? pick(record[alias]);
-      if (value === undefined) delete normalized[field];
-      else normalized[field] = value;
-    }
-    // A blank tier means "missing" (third-party); a non-string tier keeps
-    // failing validation, matching the CLI parser's type error.
-    const tier = record['tier'];
-    if (typeof tier === 'string') {
-      if (tier.trim().length === 0) delete normalized['tier'];
-      else normalized['tier'] = tier.trim();
-    }
-    // `type` trims only; the enum below rejects anything outside the CLI's
-    // plugin / managed / guide vocabulary.
-    const type = record['type'];
-    if (typeof type === 'string') normalized['type'] = type.trim();
-    // Keywords keep only non-blank strings (a junk member never fails the
-    // catalog); a non-array value reads as missing — CLI stringArrayField
-    // semantics.
-    const keywords = record['keywords'];
-    if (keywords !== undefined) {
-      const kept = Array.isArray(keywords)
-        ? keywords
-            .filter((item): item is string => typeof item === 'string')
-            .map((item) => item.trim())
-            .filter((item) => item.length > 0)
-        : [];
-      if (kept.length > 0) normalized['keywords'] = kept;
-      else delete normalized['keywords'];
-    }
-    // Version: blank or non-string reads as missing (derivation from the
-    // source then kicks in downstream) — the raw trimmed string is kept
-    // as-is otherwise (strictness lives in the update check, not here).
-    const version = pick(record['version']);
-    if (version === undefined) delete normalized['version'];
-    else normalized['version'] = version;
-    const source = pick(record['source']) ?? pick(record['url']) ?? pick(record['downloadUrl']);
-    // A source with no valid value or alias must fail validation (not slip
-    // through as whitespace): drop the key so the schema reports it missing.
-    if (source !== undefined) normalized['source'] = source;
-    else delete normalized['source'];
-    return normalized;
-  },
-  z.object({
-    id: z.string().min(1),
-    // The CLI's validateMarketplaceEntryType vocabulary (legacy aliases
-    // included); unknown types must not surface as installable plugins.
-    type: z.enum(['plugin', 'managed', 'guide']).optional(),
-    tier: z.enum(['official', 'curated']).optional(),
-    displayName: z.string().optional(),
-    description: z.string().optional(),
-    homepage: z.string().optional(),
-    keywords: z.array(z.string()).optional(),
-    version: z.string().optional(),
-    source: z.string().min(1),
-  }),
-);
-
-const rawMarketplaceSchema = z.object({
-  plugins: z.array(rawMarketplaceEntrySchema),
-});
-
-/** CLI-parity update check: both sides must be valid semver (`v` prefix and
- *  prerelease tags accepted), catalog strictly greater. */
-function semverGt(a: string, b: string): boolean {
-  return valid(a) !== null && valid(b) !== null && gt(a, b);
+function fetchWithTimeout(...args: Parameters<typeof fetch>): Promise<Response> {
+  const [input, init] = args;
+  return fetch(input, { ...init, signal: AbortSignal.timeout(MARKETPLACE_FETCH_TIMEOUT_MS) });
 }
 
 /**
- * Catalog sources may be relative to the catalog location (the production CDN
- * catalog uses `./official/*.zip`). Clients hand `source` back to
- * `POST /plugins`, whose normalizer rejects non-absolute paths — resolve
- * against the catalog URL (or, for a local catalog, its directory) so every
- * returned source is directly installable.
+ * The repo checkout's own catalog — the fallback when the default location
+ * is unreachable (offline / source-checkout dev). Absent in bundled
+ * installs, where the fallback simply never fires.
  */
-function resolveEntrySource(source: string, marketplaceUrl: string): string {
-  if (/^https?:\/\//.test(source)) return source;
-  // `file://` entry sources convert to filesystem paths up front — the
-  // install normalizer only accepts http(s) or absolute local paths.
-  if (source.startsWith('file://')) return fileURLToPath(source);
-  // Home-relative entry sources expand before any absolute/relative decision.
-  const expanded = expandHome(source);
-  if (isAbsolute(expanded)) return expanded;
-  if (/^https?:\/\//.test(marketplaceUrl)) {
-    try {
-      return new URL(expanded, marketplaceUrl).href;
-    } catch {
-      return expanded;
-    }
-  }
-  return resolve(dirname(localCatalogPath(marketplaceUrl)), expanded);
-}
-
-/** `~` / `~/` home expansion, same as the CLI loader's resolveLocalPath. */
-function expandHome(input: string): string {
-  if (input === '~') return homedir();
-  if (input.startsWith('~/')) return join(homedir(), input.slice(2));
-  return input;
-}
-
-/**
- * Derive a version from a GitHub release/tree/commit source (same shapes as
- * the CLI parser; validity follows `semver.valid`).
- */
-function deriveVersionFromGithubSource(source: string): string | undefined {
-  let url: URL;
-  try {
-    url = new URL(source);
-  } catch {
-    return undefined;
-  }
-  if (url.hostname !== 'github.com' && url.hostname !== 'www.github.com') return undefined;
-  const [, , kind, a, b] = url.pathname.split('/').filter(Boolean);
-  const ref =
-    kind === 'releases' && a === 'tag' ? b : kind === 'tree' || kind === 'commit' ? a : undefined;
-  if (ref === undefined) return undefined;
-  let decoded: string;
-  try {
-    decoded = decodeURIComponent(ref);
-  } catch {
-    decoded = ref;
-  }
-  const candidate = decoded.replace(/^v/i, '');
-  return valid(candidate) !== null ? candidate : undefined;
-}
-
-/**
- * Bare-repo GitHub sources carry no version — resolve the latest release tag
- * through the `/releases/latest` redirect (a UI route, not the rate-limited
- * API), same as the CLI. Lookups never fail the listing: any error degrades
- * to no version.
- */
-async function resolveLatestGithubRelease(
-  source: string,
-  fetchImpl: typeof fetch,
-): Promise<string | undefined> {
-  let url: URL;
-  try {
-    url = new URL(source);
-  } catch {
-    return undefined;
-  }
-  if (url.hostname !== 'github.com' && url.hostname !== 'www.github.com') return undefined;
-  // Only bare repo URLs (/<owner>/<repo>) qualify — ref tails are already
-  // handled by deriveVersionFromGithubSource.
-  const segments = url.pathname.split('/').filter(Boolean);
-  if (segments.length !== 2) return undefined;
-  const [owner, repo] = segments;
-  try {
-    const resp = await fetchImpl(`https://github.com/${owner}/${repo}/releases/latest`, {
-      redirect: 'manual',
-      signal: AbortSignal.timeout(MARKETPLACE_FETCH_TIMEOUT_MS),
-    });
-    if (resp.status !== 301 && resp.status !== 302) return undefined;
-    const location = resp.headers.get('location');
-    if (location === null) return undefined;
-    const tag = /\/releases\/tag\/([^/?#]+)/.exec(location)?.[1];
-    if (tag === undefined) return undefined;
-    let decoded: string;
-    try {
-      decoded = decodeURIComponent(tag);
-    } catch {
-      decoded = tag;
-    }
-    const candidate = decoded.replace(/^v/i, '');
-    return valid(candidate) !== null ? candidate : undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-/**
- * Local catalog location → filesystem path: `file://` conversion plus home
- * expansion.
- */
-function localCatalogPath(location: string): string {
-  return expandHome(location.startsWith('file://') ? fileURLToPath(location) : location);
-}
-
-/**
- * The repo checkout's own catalog — the CLI loader's fallback when the
- * configured catalog is unreachable (offline / source-checkout dev). Absent
- * in bundled installs, where the fallback simply never fires.
- */
-function sourceCheckoutCatalogPath(): string | undefined {
+async function getSourceCheckoutLocation(): Promise<MarketplaceLocation | undefined> {
   const candidate = resolve(import.meta.dirname, '../../../../plugins/marketplace.json');
-  return existsSync(candidate) ? candidate : undefined;
-}
-
-/**
- * Read the raw marketplace catalog JSON. Remote catalogs go through fetch;
- * local catalogs (plain path or `file://`, both accepted by the CLI loader)
- * are read from disk so the same custom catalog works for desktop/web hosts.
- * A failed remote read falls back to the source checkout's catalog when one
- * exists (CLI parity).
- */
-/**
- * Read the raw marketplace catalog JSON plus the location it was actually
- * read from — relative entry sources resolve against the latter (the
- * source-checkout fallback serves local directory sources).
- */
-async function readMarketplaceCatalog(
-  opts: PluginsRouteOptions,
-): Promise<{ raw: unknown; location: string }> {
-  const location = opts.marketplaceUrl;
-  if (!/^https?:\/\//.test(location)) {
-    return {
-      raw: JSON.parse(await readFile(localCatalogPath(location), 'utf8')),
-      location,
-    };
-  }
-  const fetchImpl = opts.fetchImpl ?? fetch;
-  try {
-    const resp = await fetchImpl(location, {
-      signal: AbortSignal.timeout(MARKETPLACE_FETCH_TIMEOUT_MS),
-    });
-    if (!resp.ok) throw new Error(`HTTP ${resp.status}`);
-    return { raw: await resp.json(), location };
-  } catch (error) {
-    const fallback =
-      opts.marketplaceIsDefault === true ? sourceCheckoutCatalogPath() : undefined;
-    if (fallback === undefined) throw error;
-    return { raw: JSON.parse(await readFile(fallback, 'utf8')), location: fallback };
-  }
+  const info = await stat(candidate).catch(() => undefined);
+  if (info?.isFile() !== true) return undefined;
+  return { raw: candidate, kind: 'local', resolved: candidate };
 }
 
 export interface PluginsRouteOptions {
@@ -357,8 +128,8 @@ export interface PluginsRouteOptions {
   /**
    * True when the catalog location is the built-in default (neither the
    * server option nor the env var set) — only then does a failed remote read
-   * fall back to the source-checkout catalog (CLI parity: an explicitly
-   * configured catalog fails hard).
+   * fall back to the source-checkout catalog and get capability markers
+   * (an explicitly configured catalog fails hard and stays unmarked).
    */
   readonly marketplaceIsDefault?: boolean;
   readonly fetchImpl?: typeof fetch;
@@ -382,9 +153,16 @@ export function registerPluginsRoutes(
       operationId: 'listPluginMarketplace',
     },
     async (req, reply) => {
-      let catalog: { raw: unknown; location: string };
+      const fetchImpl = opts.fetchImpl ?? fetchWithTimeout;
+      let read: { raw: string; location: MarketplaceLocation };
       try {
-        catalog = await readMarketplaceCatalog(opts);
+        read = await readPluginMarketplace({
+          source: opts.marketplaceUrl,
+          workDir: process.cwd(),
+          fetchImpl,
+          sourceCheckoutLocation:
+            opts.marketplaceIsDefault === true ? getSourceCheckoutLocation : undefined,
+        });
       } catch (error) {
         reply.send(
           errEnvelope(
@@ -395,31 +173,23 @@ export function registerPluginsRoutes(
         );
         return;
       }
-      const parsed = rawMarketplaceSchema.safeParse(catalog.raw);
-      if (!parsed.success) {
+      let marketplace: PluginMarketplace;
+      try {
+        marketplace = parsePluginMarketplace(read.raw, read.location);
+      } catch (error) {
         reply.send(
-          errEnvelope(ErrorCode.INTERNAL_ERROR, 'Plugin marketplace returned an invalid catalog', req.id),
+          errEnvelope(
+            ErrorCode.INTERNAL_ERROR,
+            `Plugin marketplace returned an invalid catalog: ${error instanceof Error ? error.message : String(error)}`,
+            req.id,
+          ),
         );
         return;
       }
-      const fetchImpl = opts.fetchImpl ?? fetch;
-      // Resolve sources and versions up front (parallel; latest-release
-      // lookups for bare GitHub repos ride the shared per-call timeout).
-      const resolved = await Promise.all(
-        parsed.data.plugins.map(async (entry) => {
-          const source = resolveEntrySource(entry.source, catalog.location);
-          // Entries may omit `version`: derive it from a GitHub ref tail, or
-          // look up the latest release of a bare repo source (CLI parity).
-          const version =
-            entry.version ??
-            deriveVersionFromGithubSource(source) ??
-            (await resolveLatestGithubRelease(source, fetchImpl));
-          return { entry, source, version };
-        }),
-      );
+      marketplace = await withLatestVersions(marketplace, fetchImpl);
       const installed = await core.accessor.get(IPluginService).listPlugins();
       const byId = new Map(installed.map((p) => [p.id, p]));
-      const entries: PluginMarketplaceEntryWire[] = resolved.map(({ entry, source, version }) => {
+      const entries: PluginMarketplaceEntryWire[] = marketplace.plugins.map((entry) => {
         const capabilityRow =
           opts.marketplaceIsDefault === true ? CAPABILITY_ROW_IDS[entry.id] : undefined;
         const record =
@@ -432,18 +202,17 @@ export function registerPluginsRoutes(
             ? undefined
             : { enabled: record.enabled, version: record.version };
         const updateAvailable =
-          version !== undefined &&
-          record?.version !== undefined &&
-          semverGt(version, record.version);
+          computeUpdateStatus(entry.version, record?.version, record !== undefined).kind ===
+          'update';
         return {
           id: entry.id,
           tier: entry.tier ?? 'third-party',
-          displayName: entry.displayName ?? entry.id,
+          displayName: entry.displayName,
           description: entry.description,
           homepage: entry.homepage,
-          keywords: entry.keywords,
-          version,
-          source,
+          keywords: entry.keywords === undefined ? undefined : [...entry.keywords],
+          version: entry.version,
+          source: entry.source,
           installed: installedInfo,
           updateAvailable: updateAvailable ? true : undefined,
           capabilityId: capabilityRow?.capabilityId,
